requirements.txt
transformers
torch
nltk
openai
python-dotenv
google-generativeai
